Defined Aggregate Roots Defined Entities Modified Repositories to retrict access only for Aggragates Updated PK value types from int to long To do: Create and Apply Migration with new values

+++ Library.Application/Features/LibraryFeature/Commands/CreateLibrary/CreateLibraryCommand.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace MyLibrary.Application.Features.LibraryFeature.Commands.CreateLibrary;
 
-public class CreateLibraryCommand : IRequest<int>
+public class CreateLibraryCommand : IRequest<long>
 {
     public string Name { get; set; }
     public string Description { get; set; }
